requirements.txt
absl-py==1.2.0
altair==4.2.2
antlr4-python3-runtime==4.9.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
asttokens==2.0.5
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.6.2
blis==0.7.9
cachetools==5.2.0
catalogue==2.0.8
catboost==1.2.2
certifi==2022.9.24
cffi==1.15.0
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
cmake==3.27.7
colorama==0.4.6
comm==0.1.2
comtypes==1.2.0
confection==0.0.4
contourpy==1.2.0
cryptography==36.0.2
cvzone==1.5.6
cycler==0.11.0
cymem==2.0.7
datasist==1.5.3
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.1
dlib==19.24.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
face-recognition==1.3.0
face-recognition-models==0.3.0
fake-useragent==1.3.0
fastjsonschema==2.15.3
filelock==3.12.0
filterpy==1.4.5
Flask==2.2.2
flatbuffers==23.3.3
fonttools==4.34.4
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
glob2==0.7
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.54.0
h11==0.13.0
h5py==3.7.0
hydra-core==1.3.2
idna==3.3
imageio==2.22.4
imbalanced-learn==0.10.1
importlib-metadata==6.6.0
imutils==0.5.4
inexactsearch==1.0.2
ipykernel==6.21.2
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jax==0.4.8
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.5.1
jupyter==1.0.0
jupyter-client==7.3.1
jupyter-console==6.5.1
jupyter_core==5.2.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
jupyterthemes==0.20.0
keras==2.12.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
langcodes==3.3.0
lazypredict==0.2.12
lesscpy==0.15.1
libclang==14.0.6
lightgbm==4.0.0
llvmlite==0.39.1
lxml==4.9.1
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mdurl==0.1.2
mistune==0.8.4
mizani==0.8.1
ml-dtypes==0.1.0
mpmath==1.2.1
murmurhash==1.0.9
mysql-connector-python==8.0.28
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.8
nltk==3.8.1
notebook==6.4.11
numba==0.56.4
numpy==1.23.1
oauthlib==3.2.0
omegaconf==2.3.0
opencv-python==4.7.0.72
openpyxl==3.0.10
opt-einsum==3.3.0
outcome==1.1.0
packaging==21.3
palettable==3.3.0
pandas==2.1.3
pandocfilters==1.5.0
parso==0.8.3
pathlib==1.0.1
pathy==0.10.1
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.2.0
pipline==0.0
platformdirs==3.0.0
plotly==5.14.0
plotnine==0.10.1
ply==3.11
preshed==3.0.8
prometheus-client==0.14.1
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.1
pure-eval==0.2.2
pyaml==21.10.1
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.8
pydeck==0.8.1b0
pydot==1.4.2
Pygments==2.15.1
Pympler==1.0.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
pypiwin32==223
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-version==0.0.2
pyttsx3==2.90
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pywin32==304
pywinpty==2.0.5
PyYAML==6.0
pyzmq==23.0.0
qtconsole==5.4.0
QtPy==2.3.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.3.4
rsa==4.9
rtsimple==0.9.0
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.12.0
selenium==4.1.3
Send2Trash==1.8.0
sentry-sdk==1.21.0
silpa-common==0.3
six==1.16.0
sklearn==0.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soundex==1.1.3
soupsieve==2.3.2.post1
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spellchecker==0.4
sqlparse==0.4.4
srsly==2.4.6
stack-data==0.2.0
statsmodels==0.13.2
streamlit==1.21.0
sympy==1.11.1
tenacity==8.2.2
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.12.0
tensorflow-hub==0.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.15.0
textblob==0.17.1
thinc==8.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2022.10.10
tinycss2==1.1.1
tls-client==0.2.2
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torchvision==0.15.1
tornado==6.1
tqdm==4.64.1
traitlets==5.9.0
trio==0.20.0
trio-websocket==0.9.2
typer==0.7.0
typing_extensions==4.3.0
tzdata==2022.5
tzlocal==4.3
ultralytics==8.0.88
urllib3==1.26.15
utils==1.0.1
validators==0.20.0
wasabi==1.1.1
watchdog==3.0.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
Whoosh==2.7.4
widgetsnbextension==3.6.0
wikipedia==1.4.0
wrapt==1.14.1
wsproto==1.1.0
xgboost==1.7.2
zipp==3.15.0
